+++ CHANGELOG.md
@@ -1,8 +1,12 @@
+## 0.0.3
+
+* [LanguageHelper] add `forceRebuild` to allow the plugin to force rebuild all the widgets intead of only root widgets.
+* [LanguageNotifier] add `forceRebuild`to allow the plugin to force rebuild this specific widget or not.
+
 ## 0.0.2
 
 * [Feat] Add method `LanguageHelper.instance.analyze()` to analyze the missing texts of all languages.
 * Add headers for `LanguageCodes` enum.
-
 
 ## 0.0.1+1